# Documentation/swagger.yaml
openapi: 3.0.0
paths:
...
  /auth/login:
    post:
      operationId: AuthController_login
      summary: Авторизация пользователя
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserType'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestStatusType'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedStatusType'
      tags:
        - Авторизация
  /auth/signup:
    post:
      operationId: AuthController_signup
      summary: Регистрация пользователя
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestStatusType'
      tags:
        - Авторизация
  /auth/logout:
    post:
      operationId: AuthController_logout
      summary: Выход их аккаунта
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestStatusType'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedStatusType'
      tags:
        - Авторизация
...
info:
  title: MeetupsApp
  description: Документация API
  version: 1.0.0
  contact: {}
...
servers: []
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: number
          example: '1'
        name:
          type: string
          example: Иван Петров
        phone:
          type: string
          example: 8(123)456-7890
        email:
          type: string
          example: name@example.com
        password:
          type: string
          example: password
        avatarImg:
          type: string
        role:
          type: string
          example: участник
      required:
        - id
        - name
        - email
        - password
...
    BadRequestStatusType:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        message:
          type: string
          example: Плохой запрос
      required:
        - statusCode
        - message
...
    LoginUserType:
      type: object
      properties:
        email:
          type: string
          example: name@example.com
        password:
          type: string
          example: password
      required:
        - email
        - password
    STRING:
      type: object
      properties: {}
    AuthResponse:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/STRING'
        user:
          $ref: '#/components/schemas/User'
      required:
        - status
        - user
    UnauthorizedStatusType:
      type: object
      properties:
        statusCode:
          type: number
          example: 401
        message:
          type: string
          example: Неавторизован
      required:
        - statusCode
        - message
    CreateUserDto:
      type: object
      properties:
        email:
          type: string
          example: name@example.com
        name:
          type: string
          example: Иван Петров
        password:
          type: string
          example: password
      required:
        - email
        - name
        - password
...